+++ modules/skunk/src/test/scala/composed/ComposedWorldData.scala
@@ -9,10 +9,10 @@ import cats.effect.Sync
 import cats.implicits._
 
 import edu.gemini.grackle._, skunk._
+import edu.gemini.grackle.sql.Like
 import Query._, Predicate._, Value._
 import QueryCompiler._
 import QueryInterpreter.mkErrorResult
-import SkunkPredicate._
 import _root_.skunk.codec.all._
 import cats.effect.Resource
 import _root_.skunk.Session
