+++ modules/sql/src/main/scala/FailedJoin.scala
@@ -1,0 +1,4 @@
+package edu.gemini.grackle.sql
+
+/** A sentinal value representing the empty column values from a failed join. */
+case object FailedJoin
