+++ modules/sql/src/main/scala/Row.scala
@@ -1,0 +1,6 @@
+package edu.gemini.grackle.sql
+
+/** A database row is a list of untyped values (some of which may be `FailedJoin`). */
+case class Row(elems: List[Any]) {
+  def apply(i: Int): Any = elems(i)
+}
