+++ modules/sql/src/main/scala/package.scala
@@ -1,0 +1,8 @@
+package edu.gemini.grackle
+
+package object sql {
+
+  /** A table is a list of rows. */
+  type Table = List[Row]
+
+}
